remove experimental/ and New* references
