+++ lib/playground.dart
@@ -1,14 +1,18 @@
-// Copyright (c) 2014, the Dart project authors.  Please see the AUTHORS file
+// Copyright (c) 2019, the Dart project authors.  Please see the AUTHORS file
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
 
-library playground;
+library new_playground;
 
 import 'dart:async';
-import 'dart:html' hide Document;
+import 'dart:html' hide Console;
 
+import 'package:dart_pad/editing/editor_codemirror.dart';
 import 'package:logging/logging.dart';
+import 'package:mdc_web/mdc_web.dart';
+import 'package:meta/meta.dart';
 import 'package:route_hierarchical/client.dart';
+import 'package:split/split.dart';
 
 import 'completion.dart';
 import 'context.dart';
@@ -16,11 +20,17 @@ import 'core/dependencies.dart';
 import 'core/keys.dart';
 import 'core/modules.dart';
 import 'dart_pad.dart';
-import 'dialogs.dart';
 import 'documentation.dart';
 import 'editing/editor.dart';
 import 'elements/bind.dart';
 import 'elements/elements.dart';
+import 'elements/analysis_results_controller.dart';
+import 'elements/button.dart';
+import 'elements/console.dart';
+import 'elements/counter.dart';
+import 'elements/dialog.dart';
+import 'util/keymap.dart';
+import 'elements/material_tab_controller.dart';
 import 'modules/codemirror_module.dart';
 import 'modules/dart_pad_module.dart';
 import 'modules/dartservices_module.dart';
@@ -36,6 +46,26 @@ import 'sharing/gist_storage.dart';
 import 'sharing/mutable_gist.dart';
 import 'src/ga.dart';
 import 'src/util.dart';
+import 'util/detect_flutter.dart';
+
+const codeMirrorOptions = {
+  'continueComments': {'continueLineComment': false},
+  'autofocus': false,
+  'autoCloseBrackets': true,
+  'matchBrackets': true,
+  'tabSize': 2,
+  'lineWrapping': true,
+  'indentUnit': 2,
+  'cursorHeight': 0.85,
+  'viewportMargin': 100,
+  'extraKeys': {
+    'Cmd-/': 'toggleComment',
+    'Ctrl-/': 'toggleComment',
+    'Tab': 'insertSoftTab'
+  },
+  'hintOptions': {'completeSingle': false},
+  'scrollbarStyle': 'simple',
+};
 
 Playground get playground => _playground;
 
@@ -43,117 +73,105 @@ Playground _playground;
 
 final Logger _logger = Logger('dartpad');
 
-/// Controls whether we request compilation using dart2js or DDC.
-const bool _useDDC = false;
-
 void init() {
   _playground = Playground();
 }
 
 class Playground implements GistContainer, GistController {
-  DivElement get _editpanel => querySelector('#editpanel');
+  final MutableGist editableGist = MutableGist(Gist());
+  final GistStorage _gistStorage = GistStorage();
+  MDCButton newButton;
+  MDCButton resetButton;
+  MDCButton formatButton;
+  MDCButton samplesButton;
+  MDCButton runButton;
+  MDCButton editorConsoleTab;
+  MDCButton editorDocsTab;
+  MDCButton closePanelButton;
+  MDCButton moreMenuButton;
+  DElement editorPanelHeader;
+  DElement editorPanelFooter;
+  MDCMenu samplesMenu;
+  MDCMenu moreMenu;
+  Dialog dialog;
+  NewPadDialog newPadDialog;
+  DElement titleElement;
+  MaterialTabController webLayoutTabController;
+  DElement webTabBar;
+  DElement webOutputLabel;
 
-  DivElement get _outputpanel => querySelector('#output');
+  Splitter splitter;
+  Splitter rightSplitter;
+  bool rightSplitterConfigured = false;
+  TabExpandController tabExpandController;
+  AnalysisResultsController analysisResultsController;
 
-  IFrameElement get _frame => querySelector('#frame');
-
-  bool get _isCompletionActive => editor.completionActive;
-
-  DivElement get _docPanel => querySelector('#documentation');
-
-  DButton runButton;
-  DButton formatButton;
-  DOverlay overlay;
   DBusyLight busyLight;
   DBusyLight consoleBusyLight;
+
   Editor editor;
   PlaygroundContext _context;
   Future _analysisRequest;
-  MutableGist editableGist = MutableGist(Gist());
-  final _gistStorage = GistStorage();
-  DContentEditable titleEditable;
+  Layout _layout;
 
-  TabController sourceTabController;
-  TabController outputTabController;
-  SharingDialog sharingDialog;
-  KeysDialog settings;
-
-  // We store the last returned shared gist; it's used to update the url.
+  // The last returned shared gist used to update the url.
   Gist _overrideNextRouteGist;
   DocHandler docHandler;
 
+  Console _leftConsole;
+  Console _rightConsole;
+  Counter unreadConsoleCounter;
+
   Playground() {
-    sourceTabController = TabController();
-    for (String name in ['dart', 'html', 'css']) {
-      sourceTabController.registerTab(
-          TabElement(querySelector('#${name}tab'), name: name, onSelect: () {
-        Element issuesElement = querySelector('#issues');
-        issuesElement.style.display = name == 'dart' ? 'block' : 'none';
-        ga.sendEvent('edit', name);
-        _context.switchTo(name);
-      }));
-    }
-
-    overlay = DOverlay(querySelector('#frame-overlay'));
-
-    sharingDialog = SharingDialog(this, this);
-
-    DButton newButton = DButton(querySelector('#newbutton'));
-    OkCancelDialog newDialog = OkCancelDialog(
-        'Create New Pad', 'Discard changes to the current pad?', createNewGist,
-        okText: 'Discard');
-    newButton.onClick.listen((_) {
-      newDialog.show();
+    _initModules().then((_) {
+      _initPlayground();
+      _initDialogs();
+      _initBusyLights();
+      _initGistNameHeader();
+      _initGistStorage();
+      _initLayoutDetection();
+      _initButtons();
+      _initLabels();
+      _initSamplesMenu();
+      _initMoreMenu();
+      _initSplitters();
+      _initTabs();
+      _initLayout();
+      _initConsoles();
     });
+  }
 
-    DButton resetButton = DButton(querySelector('#resetbutton'));
-    OkCancelDialog resetDialog = OkCancelDialog(
-        'Reset Pad', 'Discard changes to the current pad?', _resetGists,
-        okText: 'Discard', cancelText: 'Cancel');
-    resetButton.onClick.listen((_) {
-      resetDialog.show();
-    });
+  DivElement get _editorHost => querySelector('#editor-host');
+  DivElement get _rightConsoleElement => querySelector('#right-output-panel');
+  DivElement get _rightConsoleContentElement =>
+      querySelector('#right-output-panel-content');
+  DivElement get _leftConsoleElement => querySelector('#left-output-panel');
+  IFrameElement get _frame => querySelector('#frame');
+  DivElement get _rightDocPanel => querySelector('#right-doc-panel');
+  DivElement get _rightDocContentElement =>
+      querySelector('#right-doc-panel-content');
+  DivElement get _leftDocPanel => querySelector('#left-doc-panel');
+  DivElement get _editorPanelHeader => querySelector('#editor-panel-header');
+  DivElement get _editorPanelFooter => querySelector('#editor-panel-footer');
+  bool get _isCompletionActive => editor.completionActive;
 
-    editableGist.onDirtyChanged.listen((val) {
-      resetButton.disabled = !val;
-    });
+  void _initDialogs() {
+    dialog = Dialog();
+    newPadDialog = NewPadDialog();
+  }
 
-    DButton shareButton = DButton(querySelector('#sharebutton'));
-
-    shareButton.onClick.listen((Event e) => window.open(
-        'https://github.com/dart-lang/dart-pad/wiki/Sharing-Guide',
-        '_sharing'));
-
-    // Sharing is currently disabled pending establishing OAuth2 configurations with Github.
-    // shareButton.onClick.listen((Event e) => _createSummary()
-    //    .then((GistSummary summary) => sharingDialog.showWithSummary(summary)));
-
-    formatButton = DButton(querySelector('#formatbutton'));
-    formatButton.onClick.listen((Event e) => _format());
-
-    runButton = DButton(querySelector('#runbutton'));
-    runButton.onClick.listen((e) {
-      _handleRun();
-
-      // On a mobile device, focusing the editing area causes the keyboard to
-      // pop up when the user hits the run button.
-      if (!isMobile()) _context.focus();
-    });
-
-    // Listen for the keyboard button.
-    querySelector('#keyboard-button').onClick.listen((_) => settings.show());
-
+  void _initBusyLights() {
     busyLight = DBusyLight(querySelector('#dartbusy'));
     consoleBusyLight = DBusyLight(querySelector('#consolebusy'));
+  }
 
+  void _initGistNameHeader() {
     // Update the title on changes.
-    titleEditable = DContentEditable(querySelector('header .header-gist-name'));
-    bind(titleEditable.onChanged, editableGist.property('description'));
-    bind(editableGist.property('description'), titleEditable.textProperty);
-    editableGist.onDirtyChanged.listen((val) {
-      titleEditable.element.classes.toggle('dirty', val);
-    });
+    titleElement = DElement(querySelector('header .header-gist-name'));
+    bind(editableGist.property('description'), titleElement.textProperty);
+  }
 
+  void _initGistStorage() {
     // If there was a change, and the gist is dirty, write the gist's contents
     // to storage.
     debounceStream(mutableGist.onChanged, Duration(milliseconds: 100))
@@ -162,209 +180,221 @@ class Playground implements GistContainer, GistController {
         _gistStorage.setStoredGist(mutableGist.createGist());
       }
     });
+  }
 
-    SelectElement select = querySelector('#samples');
-    select.onChange.listen((_) => _handleSelectChanged(select));
-
-    // Show the about box on title clicks.
-    querySelector('div.header-title').onClick.listen((e) {
-      e.preventDefault();
-      _showAboutBox();
-    });
-
-    // Show the about box on version text clicks.
-    querySelector('#dartpad_version').onClick.listen((e) {
-      e.preventDefault();
-      _showAboutBox();
-    });
-
-    _initModules().then((_) {
-      _initPlayground();
+  void _initLayoutDetection() {
+    debounceStream(mutableGist.onChanged, Duration(milliseconds: 32))
+        .listen((_) {
+      if (hasFlutterContent(_context.dartSource)) {
+        _changeLayout(Layout.flutter);
+      } else if (hasHtmlContent(_context.dartSource)) {
+        _changeLayout(Layout.html);
+      } else {
+        _changeLayout(Layout.dart);
+      }
     });
   }
 
-  void _showAboutBox() {
-    dartServices
-        .version()
-        .timeout(Duration(seconds: 2))
-        .then((VersionResponse ver) {
-      print('Dart SDK version ${ver.sdkVersion} (${ver.sdkVersionFull})');
-      print('CodeMirror: ${CodeMirrorModule.version}');
-      AboutDialog(ver.sdkVersionFull)..show();
-    }).catchError((e) {
-      AboutDialog()..show();
-    });
-  }
-
-  void _resetGists() {
-    _gistStorage.clearStoredGist();
-    editableGist.reset();
-    // Delay to give time for the model change event to propagate through
-    // to the editor component (which is where `_performAnalysis()` pulls
-    // the Dart source from).
-    Timer.run(_performAnalysis);
-    _clearOutput();
-  }
-
-  Future showHome(RouteEnterEvent event) async {
-    // Don't auto-run if we're re-loading some unsaved edits; the gist might
-    // have halting issues (#384).
-    bool loadedFromSaved = false;
-    Uri url = Uri.parse(window.location.toString());
-
-    if (url.hasQuery &&
-        url.queryParameters['id'] != null &&
-        isLegalGistId(url.queryParameters['id'])) {
-      _showGist(url.queryParameters['id']);
-    } else if (url.hasQuery && url.queryParameters['export'] != null) {
-      UuidContainer requestId = UuidContainer()
-        ..uuid = url.queryParameters['export'];
-      Future<PadSaveObject> exportPad =
-          dartSupportServices.pullExportContent(requestId);
-      await exportPad.then((pad) {
-        Gist blankGist = createSampleDartGist();
-        blankGist.getFile('main.dart').content = pad.dart;
-        blankGist.getFile('index.html').content = pad.html;
-        blankGist.getFile('styles.css').content = pad.css;
-        editableGist.setBackingGist(blankGist);
+  void _initButtons() {
+    newButton = MDCButton(querySelector('#new-button'))
+      ..onClick.listen((_) => _showCreateGistDialog());
+    resetButton = MDCButton(querySelector('#reset-button'))
+      ..onClick.listen((_) => _showResetDialog());
+    formatButton = MDCButton(querySelector('#format-button'))
+      ..onClick.listen((_) => _format());
+    samplesButton = MDCButton(querySelector('#samples-dropdown-button'))
+      ..onClick.listen((e) {
+        samplesMenu.open = !samplesMenu.open;
       });
-    } else if (url.hasQuery && url.queryParameters['source'] != null) {
-      UuidContainer gistId = await dartSupportServices.retrieveGist(
-          id: url.queryParameters['source']);
-      Gist backing;
 
-      try {
-        backing = await gistLoader.loadGist(gistId.uuid);
-      } catch (ex) {
-        print(ex);
-        backing = Gist();
-      }
+    runButton = MDCButton(querySelector('#run-button'))
+      ..onClick.listen((_) {
+        _handleRun();
+      });
+    editorConsoleTab = MDCButton(querySelector('#editor-panel-console-tab'));
+    editorDocsTab = MDCButton(querySelector('#editor-panel-docs-tab'));
+    closePanelButton =
+        MDCButton(querySelector('#editor-panel-close-button'), isIcon: true);
+    moreMenuButton = MDCButton(querySelector('#more-menu-button'), isIcon: true)
+      ..onClick.listen((_) {
+        moreMenu.open = !moreMenu.open;
+      });
+    querySelector('#keyboard-button')
+        .onClick
+        .listen((_) => _showKeyboardDialog());
+  }
 
-      editableGist.setBackingGist(backing);
-      await router.go('gist', {'gist': backing.id});
-    } else if (_gistStorage.hasStoredGist && _gistStorage.storedId == null) {
-      loadedFromSaved = true;
+  void _initLabels() {
+    var webOutputLabelElement = querySelector('#web-output-label');
+    if (webOutputLabelElement != null) {
+      webOutputLabel = DElement(webOutputLabelElement);
+    }
+  }
 
-      Gist blankGist = Gist();
-      editableGist.setBackingGist(blankGist);
+  void _initSamplesMenu() {
+    var element = querySelector('#samples-menu');
 
-      Gist storedGist = _gistStorage.getStoredGist();
-      editableGist.description = storedGist.description;
-      for (GistFile file in storedGist.files) {
-        editableGist.getGistFile(file.name).content = file.content;
-      }
-    } else {
-      editableGist.setBackingGist(createSampleDartGist());
+    var samples = [
+      Sample('215ba63265350c02dfbd586dfd30b8c3', 'Hello World', Layout.dart),
+      Sample('e93b969fed77325db0b848a85f1cf78e', 'Int to Double', Layout.dart),
+      Sample('b60dc2fc7ea49acecb1fd2b57bf9be57', 'Mixins', Layout.dart),
+      Sample('7d78af42d7b0aedfd92f00899f93561b', 'Fibonacci', Layout.dart),
+      Sample('a559420eed617dab7a196b5ea0b64fba', 'Sunflower', Layout.html),
+      Sample('cb9b199b1085873de191e32a1dd5ca4f', 'WebSockets', Layout.html),
+      Sample('b6409e10de32b280b8938aa75364fa7b', 'Counter', Layout.flutter),
+      Sample('1ae68b761529fd721b84d8709c6017bd', 'Todo App', Layout.flutter),
+      Sample(
+          'ceb5101e540378761391e2b7f6b64325', 'Sliding Square', Layout.flutter),
+    ];
+
+    var listElement = UListElement()
+      ..classes.add('mdc-list')
+      ..attributes.addAll({
+        'aria-hidden': 'true',
+        'aria-orientation': 'vertical',
+        'tabindex': '-1'
+      });
+
+    element.children.add(listElement);
+
+    // Helper function to create LIElement with correct attributes and classes
+    // for material-components-web
+    LIElement _menuElement(Sample sample) {
+      return LIElement()
+        ..classes.add('mdc-list-item')
+        ..attributes.addAll({'role': 'menuitem'})
+        ..children.add(
+          ImageElement()
+            ..classes.add('mdc-list-item__graphic')
+            ..src = 'pictures/logo_${_layoutToString(sample.layout)}.png',
+        )
+        ..children.add(
+          SpanElement()
+            ..classes.add('mdc-list-item__text')
+            ..text = sample.name,
+        );
     }
 
-    _clearOutput();
-    // We delay this because of the latency in populating the editors from the
-    // gist data.
-    Timer.run(_autoSwitchSourceTab);
+    for (var sample in samples) {
+      listElement.children.add(_menuElement(sample));
+    }
 
-    // Analyze and run it.
-    Timer.run(() {
-      _performAnalysis().then((bool result) {
-        // Only auto-run if the static analysis comes back clean.
-        if (result && !loadedFromSaved) _handleRun();
-        if (url.hasQuery && url.queryParameters['line'] != null) {
-          _jumpToLine(int.parse(url.queryParameters['line']));
-        }
-      }).catchError((e) => null);
+    samplesMenu = MDCMenu(element)
+      ..setAnchorCorner(AnchorCorner.bottomLeft)
+      ..setAnchorElement(querySelector('#samples-dropdown-button'))
+      ..hoistMenuToBody();
+
+    samplesMenu.listen('MDCMenu:selected', (e) {
+      var index = (e as CustomEvent).detail['index'];
+      var gistId = samples.elementAt(index).gistId;
+      router.go('gist', {'gist': gistId});
     });
   }
 
-  void showGist(RouteEnterEvent event) {
-    String gistId = event.parameters['gist'];
+  void _initMoreMenu() {
+    moreMenu = MDCMenu(querySelector('#more-menu'))
+      ..setAnchorCorner(AnchorCorner.bottomLeft)
+      ..setAnchorElement(querySelector('#more-menu-button'))
+      ..hoistMenuToBody();
+    moreMenu.listen('MDCMenu:selected', (e) {
+      var idx = (e as CustomEvent).detail['index'];
+      switch (idx) {
+        case 0:
+          _showSharingPage();
+          break;
+        case 1:
+          _showGitHubPage();
+          break;
+      }
+    });
+  }
 
-    _clearOutput();
+  void _initSplitters() {
+    var editorPanel = querySelector('#editor-panel');
+    var outputPanel = querySelector('#output-panel');
 
-    if (!isLegalGistId(gistId)) {
-      showHome(event);
+    splitter = flexSplit(
+      [editorPanel, outputPanel],
+      horizontal: true,
+      gutterSize: 6,
+      sizes: [50, 50],
+      minSize: [100, 100],
+    );
+  }
+
+  void _initRightSplitter() {
+    if (rightSplitterConfigured) {
       return;
     }
 
-    _showGist(gistId);
+    var outputHost = querySelector('#right-output-panel');
+    rightSplitter = flexSplit(
+      [outputHost, _rightDocPanel],
+      horizontal: false,
+      gutterSize: 6,
+      sizes: [50, 50],
+      minSize: [100, 100],
+    );
+    rightSplitterConfigured = true;
   }
 
-  // GistContainer interface
-  @override
-  MutableGist get mutableGist => editableGist;
-
-  @override
-  void overrideNextRoute(Gist gist) {
-    _overrideNextRouteGist = gist;
+  void _disposeRightSplitter() {
+    if (!rightSplitterConfigured) {
+      // The right splitter might already be destroyed.
+      return;
+    }
+    rightSplitter?.destroy();
+    rightSplitterConfigured = false;
   }
 
-  @override
-  Future createNewGist() {
-    _gistStorage.clearStoredGist();
-
-    if (ga != null) ga.sendEvent('main', 'new');
-
-    DToast.showMessage('New pad created');
-    router.go('gist', {'gist': ''}, forceReload: true);
-
-    return Future.value();
-  }
-
-  void _showGist(String gistId) {
-    // Don't auto-run if we're re-loading some unsaved edits; the gist might
-    // have halting issues (#384).
-    bool loadedFromSaved = false;
-
-    // When sharing, we have to pipe the returned (created) gist through the
-    // routing library to update the url properly.
-    if (_overrideNextRouteGist != null && _overrideNextRouteGist.id == gistId) {
-      editableGist.setBackingGist(_overrideNextRouteGist);
-      _overrideNextRouteGist = null;
+  void _initOutputPanelTabs() {
+    if (tabExpandController != null) {
       return;
     }
 
-    _overrideNextRouteGist = null;
-
-    gistLoader.loadGist(gistId).then((Gist gist) {
-      editableGist.setBackingGist(gist);
-
-      if (_gistStorage.hasStoredGist && _gistStorage.storedId == gistId) {
-        loadedFromSaved = true;
-
-        Gist storedGist = _gistStorage.getStoredGist();
-        editableGist.description = storedGist.description;
-        for (GistFile file in storedGist.files) {
-          editableGist.getGistFile(file.name).content = file.content;
-        }
-      }
-
-      _clearOutput();
-
-      // We delay this because of the latency in populating the editors from the
-      // gist data.
-      Timer.run(_autoSwitchSourceTab);
-
-      if (isWebContentShowing != gist.hasWebContent()) {
-        if (gist.hasWebContent()) {
-          showWebContent();
-        } else {
-          hideWebContent();
-        }
-      }
-
-      // Analyze and run it.
-      Timer.run(() {
-        _performAnalysis().then((bool result) {
-          // Only auto-run if the static analysis comes back clean.
-          if (result && !loadedFromSaved) _handleRun();
-        }).catchError((e) => null);
-      });
-    }).catchError((e) {
-      String message = 'Error loading gist $gistId.';
-      DToast.showMessage(message);
-      _logger.severe('$message: $e');
-    });
+    tabExpandController = TabExpandController(
+      consoleButton: editorConsoleTab,
+      docsButton: editorDocsTab,
+      closeButton: closePanelButton,
+      docsElement: _leftDocPanel,
+      consoleElement: _leftConsoleElement,
+      topSplit: _editorHost,
+      bottomSplit: _editorPanelFooter,
+      unreadCounter: unreadConsoleCounter,
+    );
   }
 
-  Future _initModules() {
+  void _disposeOutputPanelTabs() {
+    tabExpandController?.dispose();
+    tabExpandController = null;
+  }
+
+  void _initTabs() {
+    webTabBar = DElement(querySelector('#web-tab-bar'));
+    webLayoutTabController =
+        MaterialTabController(MDCTabBar(webTabBar.element));
+    for (String name in ['dart', 'html', 'css']) {
+      webLayoutTabController.registerTab(
+          TabElement(querySelector('#$name-tab'), name: name, onSelect: () {
+        ga.sendEvent('edit', name);
+        _context.switchTo(name);
+      }));
+    }
+  }
+
+  void _initLayout() {
+    editorPanelHeader = DElement(_editorPanelHeader);
+    editorPanelFooter = DElement(_editorPanelFooter);
+    _changeLayout(Layout.dart);
+  }
+
+  void _initConsoles() {
+    _leftConsole = Console(DElement(_leftConsoleElement));
+    _rightConsole = Console(DElement(_rightConsoleContentElement));
+    unreadConsoleCounter = Counter(querySelector('#unread-console-counter'));
+  }
+
+  Future<void> _initModules() async {
     ModuleManager modules = ModuleManager();
 
     modules.register(DartPadModule());
@@ -372,36 +402,10 @@ class Playground implements GistContainer, GistController {
     modules.register(DartSupportServicesModule());
     modules.register(CodeMirrorModule());
 
-    return modules.start();
+    await modules.start();
   }
 
   void _initPlayground() {
-    var disablePointerEvents = () {
-      _frame.style.pointerEvents = 'none';
-    };
-    var enablePointerEvents = () {
-      _frame.style.pointerEvents = 'inherit';
-    };
-
-    DSplitter editorSplitter = DSplitter(querySelector('#editor_split'),
-        onDragStart: disablePointerEvents, onDragEnd: enablePointerEvents);
-    editorSplitter.onPositionChanged.listen((pos) {
-      state['editor_split'] = pos;
-      editor.resize();
-    });
-    if (state['editor_split'] != null) {
-      editorSplitter.position = state['editor_split'];
-    }
-
-    DSplitter outputSplitter = DSplitter(querySelector('#output_split'),
-        onDragStart: disablePointerEvents, onDragEnd: enablePointerEvents);
-    outputSplitter.onPositionChanged.listen((pos) {
-      state['output_split'] = pos;
-    });
-    if (state['output_split'] != null) {
-      outputSplitter.position = state['output_split'];
-    }
-
     // Set up the iframe.
     deps[ExecutionService] = ExecutionServiceIFrame(_frame);
     executionService.onStdout.listen(_showOutput);
@@ -413,16 +417,19 @@ class Playground implements GistContainer, GistController {
     // Set up the gist loader.
     deps[GistLoader] = GistLoader.defaultFilters();
 
-    // Set up the editing area.
-    editor = editorFactory.createFromElement(_editpanel);
-    _editpanel.children.first.attributes['flex'] = '';
-    editor.resize();
+    // Set up CodeMirror
+    editor = (editorFactory as CodeMirrorFactory)
+        .createFromElement(_editorHost, options: codeMirrorOptions)
+          ..theme = 'darkpad'
+          ..mode = 'dart';
 
+    // set up key bindings
     keys.bind(['ctrl-s'], _handleSave, 'Save', hidden: true);
     keys.bind(['ctrl-enter'], _handleRun, 'Run');
     keys.bind(['f1'], () {
       ga.sendEvent('main', 'help');
-      docHandler.generateDoc(_docPanel);
+      docHandler.generateDoc(_rightDocContentElement);
+      docHandler.generateDoc(_leftDocPanel);
     }, 'Documentation');
 
     keys.bind(['alt-enter'], () {
@@ -434,104 +441,30 @@ class Playground implements GistContainer, GistController {
     }, 'Completion');
 
     keys.bind(['shift-ctrl-/', 'shift-macctrl-/'], () {
-      if (settings.isShowing) {
-        settings.hide();
-      } else {
-        settings.show();
-      }
+      _showKeyboardDialog();
     }, 'Keyboard Shortcuts');
-
-    settings = KeysDialog(keys.inverseBindings);
+    keys.bind(['shift-ctrl-f', 'shift-macctrl-f'], () {
+      _format();
+    }, 'Format');
 
     document.onKeyUp.listen((e) {
       if (editor.completionActive ||
           DocHandler.cursorKeys.contains(e.keyCode)) {
-        docHandler.generateDoc(_docPanel);
+        docHandler.generateDoc(_rightDocContentElement);
+        docHandler.generateDoc(_leftDocPanel);
       }
       _handleAutoCompletion(e);
     });
 
-    outputTabController = TabController()
-      ..registerTab(
-          TabElement(querySelector('#resulttab'), name: 'result', onSelect: () {
-        ga.sendEvent('view', 'result');
-        querySelector('#frame').style.visibility = 'visible';
-        querySelector('#output').style.visibility = 'hidden';
-      }))
-      ..registerTab(TabElement(querySelector('#consoletab'), name: 'console',
-          onSelect: () {
-        ga.sendEvent('view', 'console');
-        querySelector('#output').style.visibility = 'visible';
-        querySelector('#frame').style.visibility = 'hidden';
-      }));
-
     _context = PlaygroundContext(editor);
-    _context.onModeChange.listen((_) {
-      formatButton.disabled = _context.activeMode != 'dart';
-    });
     deps[Context] = _context;
 
     editorFactory.registerCompleter(
         'dart', DartCompleter(dartServices, _context.dartDocument));
 
-    _context.onHtmlDirty.listen((_) => busyLight.on());
-    _context.onHtmlReconcile.listen((_) {
-      executionService.replaceHtml(_context.htmlSource);
-      busyLight.reset();
-    });
-
-    _context.onCssDirty.listen((_) => busyLight.on());
-    _context.onCssReconcile.listen((_) {
-      executionService.replaceCss(_context.cssSource);
-      busyLight.reset();
-    });
-
     _context.onDartDirty.listen((_) => busyLight.on());
     _context.onDartReconcile.listen((_) => _performAnalysis());
 
-    // Listen for changes that would effect the documentation panel.
-    editor.onMouseDown.listen((e) {
-      // Delay to give codemirror time to process the mouse event.
-      Timer.run(() {
-        if (!_context.cursorPositionIsWhitespace()) {
-          docHandler.generateDoc(_docPanel);
-        }
-      });
-    });
-    context.onModeChange.listen((_) => docHandler.generateDoc(_docPanel));
-
-    // Listen for clicks on the 'Show web content' checkbox.
-    showWebContentInputElement.onClick.listen((MouseEvent event) {
-      event.preventDefault();
-
-      // Delay a bit as it looks like we can't programmatically change the value
-      // of a checkbox in an event handler.
-      Timer(const Duration(milliseconds: 100), () {
-        final bool isChecked = showWebContentInputElement.checked;
-
-        if (isChecked) {
-          if (_context.hasWebContent()) {
-            final OkCancelDialog dialog = OkCancelDialog(
-              'Hide web content',
-              'Discard the contents of the HTML and CSS tabs?',
-              () {
-                _context.htmlSource = '';
-                _context.cssSource = '';
-                hideWebContent();
-              },
-              okText: 'Discard',
-            );
-            dialog.show();
-          } else {
-            hideWebContent();
-          }
-        } else {
-          showWebContent();
-        }
-      });
-    });
-
-    // Bind the editable files to the gist.
     Property htmlFile =
         GistFileProperty(editableGist.getGistFile('index.html'));
     Property htmlDoc = EditorDocumentProperty(_context.htmlDocument, 'html');
@@ -548,19 +481,53 @@ class Playground implements GistContainer, GistController {
     bind(dartDoc, dartFile);
     bind(dartFile, dartDoc);
 
+    // Listen for changes that would effect the documentation panel.
+    editor.onMouseDown.listen((e) {
+      // Delay to give codemirror time to process the mouse event.
+      Timer.run(() {
+        if (!_context.cursorPositionIsWhitespace()) {
+          docHandler.generateDoc(_rightDocContentElement);
+          docHandler.generateDoc(_leftDocPanel);
+        }
+      });
+    });
+
     // Set up the router.
     deps[Router] = Router();
     router.root.addRoute(name: 'home', defaultRoute: true, enter: showHome);
+    router.root.addRoute(
+        name: 'dart',
+        path: '/dart',
+        defaultRoute: false,
+        enter: (_) => showNew(Layout.dart));
+    router.root.addRoute(
+        name: 'html',
+        path: '/html',
+        defaultRoute: false,
+        enter: (_) => showNew(Layout.html));
+    router.root.addRoute(
+        name: 'flutter',
+        path: '/flutter',
+        defaultRoute: false,
+        enter: (_) => showNew(Layout.flutter));
     router.root.addRoute(name: 'gist', path: '/:gist', enter: showGist);
     router.listen();
 
     docHandler = DocHandler(editor, _context);
 
     dartServices.version().then((VersionResponse version) {
-      // "Based on Dart SDK 1.25.0-dev"
+      // "Based on Dart SDK 2.4.0"
       String versionText = 'Based on Dart SDK ${version.sdkVersionFull}';
-      querySelector('#dartpad_version').text = versionText;
+      querySelector('#dartpad-version').text = versionText;
     }).catchError((e) => null);
+
+    analysisResultsController = AnalysisResultsController(
+        DElement(querySelector('#issues')),
+        DElement(querySelector('#issues-message')),
+        DElement(querySelector('#issues-toggle')))
+      ..onIssueClick.listen((issue) {
+        _jumpTo(issue.line, issue.charStart, issue.charLength, focus: true);
+      });
 
     _finishedInit();
   }
@@ -593,10 +560,167 @@ class Playground implements GistContainer, GistController {
     }
   }
 
+  Future<void> showNew(Layout layout) async {
+    var loadResult = _loadGist();
+    var shouldAutoRun = loadResult == LoadGistResult.storage;
+
+    // If no gist was loaded, use a new Dart gist.
+    if (loadResult == LoadGistResult.none) {
+      editableGist.setBackingGist(_createGist(layout));
+
+      // Store the gist so that the same sample is loaded when the page is
+      // refreshed.
+      _gistStorage.setStoredGist(editableGist.createGist());
+
+      _changeLayout(layout);
+    } else {
+      // If a Gist was loaded from storage or from a Gist, use the layout
+      // detected by reading the code.
+      _changeLayout(_detectLayout(editableGist.backingGist));
+    }
+
+    // Clear console output and update the layout if necessary.
+    var url = Uri.parse(window.location.toString());
+    _clearOutput();
+
+    if (url.hasQuery && url.queryParameters['line'] != null) {
+      _jumpToLine(int.parse(url.queryParameters['line']));
+    }
+
+    await _analyzeAndRun(autoRun: shouldAutoRun);
+  }
+
+  Gist _createGist(Layout layout) {
+    switch (layout) {
+      case Layout.flutter:
+        return createSampleFlutterGist();
+      case Layout.html:
+        return createSampleHtmlGist();
+      default:
+        return createSampleDartGist();
+    }
+  }
+
+  Future<void> showHome(RouteEnterEvent event) async {
+    await showNew(Layout.dart);
+  }
+
+  /// Loads the gist provided by the 'id' query parameter or stored in
+  /// [GistStorage].
+  LoadGistResult _loadGist() {
+    Uri url = Uri.parse(window.location.toString());
+
+    if (url.hasQuery &&
+        url.queryParameters['id'] != null &&
+        isLegalGistId(url.queryParameters['id'])) {
+      _showGist(url.queryParameters['id']);
+      return LoadGistResult.queryParameter;
+    }
+
+    if (_gistStorage.hasStoredGist && _gistStorage.storedId == null) {
+      Gist blankGist = Gist();
+      editableGist.setBackingGist(blankGist);
+
+      Gist storedGist = _gistStorage.getStoredGist();
+
+      // Set the editable gist's backing gist so that the route handler can
+      // detect the project type.
+      editableGist.setBackingGist(storedGist);
+
+      editableGist.description = storedGist.description;
+      for (GistFile file in storedGist.files) {
+        editableGist.getGistFile(file.name).content = file.content;
+      }
+      return LoadGistResult.storage;
+    }
+
+    return LoadGistResult.none;
+  }
+
+  /// Analyzes and runs the gist.  Auto-runs the gist if [autoRun] is true and
+  /// the analyzer comes back clean.
+  Future<void> _analyzeAndRun({bool autoRun = false}) {
+    var completer = Completer();
+    Timer.run(() async {
+      try {
+        var result = await _performAnalysis();
+        if (result && !autoRun) {
+          _handleRun();
+        }
+      } catch (e) {
+        // ignore errors
+      }
+      completer.complete();
+    });
+    return completer.future;
+  }
+
+  void showGist(RouteEnterEvent event) {
+    String gistId = event.parameters['gist'];
+
+    _clearOutput();
+
+    if (!isLegalGistId(gistId)) {
+      showHome(event);
+      return;
+    }
+
+    _showGist(gistId);
+  }
+
+  void _showGist(String gistId) {
+    // Don't auto-run if we're re-loading some unsaved edits; the gist might
+    // have halting issues (#384).
+    bool loadedFromSaved = false;
+
+    // When sharing, we have to pipe the returned (created) gist through the
+    // routing library to update the url properly.
+    if (_overrideNextRouteGist != null && _overrideNextRouteGist.id == gistId) {
+      editableGist.setBackingGist(_overrideNextRouteGist);
+      _overrideNextRouteGist = null;
+      return;
+    }
+
+    _overrideNextRouteGist = null;
+
+    gistLoader.loadGist(gistId).then((Gist gist) {
+      editableGist.setBackingGist(gist);
+
+      if (_gistStorage.hasStoredGist && _gistStorage.storedId == gistId) {
+        loadedFromSaved = true;
+
+        Gist storedGist = _gistStorage.getStoredGist();
+        editableGist.description = storedGist.description;
+        for (GistFile file in storedGist.files) {
+          editableGist.getGistFile(file.name).content = file.content;
+        }
+      }
+
+      _clearOutput();
+
+      _changeLayout(_detectLayout(gist));
+
+      // Analyze and run it.
+      Timer.run(() {
+        _performAnalysis().then((bool result) {
+          // Only auto-run if the static analysis comes back clean.
+          if (result && !loadedFromSaved) _handleRun();
+        }).catchError((e) => null);
+      });
+    }).catchError((e) {
+      String message = 'Error loading gist $gistId.';
+      _showSnackbar(message);
+      _logger.severe('$message: $e');
+    });
+  }
+
+  void _showKeyboardDialog() {
+    dialog.showOk('Keyboard shortcuts', keyMapToHtml(keys.inverseBindings));
+  }
+
   void _handleRun() async {
     ga.sendEvent('main', 'run');
     runButton.disabled = true;
-    overlay.visible = true;
 
     Stopwatch compilationTimer = Stopwatch()..start();
 
@@ -604,7 +728,7 @@ class Playground implements GistContainer, GistController {
       ..source = context.dartSource;
 
     try {
-      if (_useDDC) {
+      if (hasFlutterContent(_context.dartSource)) {
         final CompileDDCResponse response = await dartServices
             .compileDDC(compileRequest)
             .timeout(longServiceCallTimeout);
@@ -615,7 +739,6 @@ class Playground implements GistContainer, GistController {
           compilationTimer.elapsedMilliseconds,
         );
 
-        _autoSwitchOutputTab();
         _clearOutput();
 
         return executionService.execute(
@@ -635,10 +758,9 @@ class Playground implements GistContainer, GistController {
           compilationTimer.elapsedMilliseconds,
         );
 
-        _autoSwitchOutputTab();
         _clearOutput();
 
-        return executionService.execute(
+        return await executionService.execute(
           _context.htmlSource,
           _context.cssSource,
           response.result,
@@ -647,67 +769,14 @@ class Playground implements GistContainer, GistController {
     } catch (e) {
       ga.sendException('${e.runtimeType}');
       final message = (e is DetailedApiRequestError) ? e.message : '$e';
-      DToast.showMessage('Error compiling to JavaScript');
+      _showSnackbar('Error compiling to JavaScript');
+      _clearOutput();
       _showOutput('Error compiling to JavaScript:\n$message', error: true);
     } finally {
       runButton.disabled = false;
-      overlay.visible = false;
+      webOutputLabel.setAttr('hidden');
     }
   }
-
-  /// Switch to the Dart / Html / Css tab depending on the sample type.
-  void _autoSwitchSourceTab() {
-    String htmlSrc = _context.htmlSource.trim();
-    String dartSrc = _context.dartSource.trim();
-
-    if (htmlSrc.isEmpty && dartSrc.isNotEmpty) {
-      sourceTabController.selectTab('dart');
-    }
-  }
-
-  /// Switch to the console or html results tab depending on whether the sample
-  /// has html content or not.
-  void _autoSwitchOutputTab() {
-    if (!isWebContentShowing) {
-      return;
-    }
-
-    String htmlSrc = _context.htmlSource.trim();
-    String dartSrc = _context.dartSource.trim();
-
-    final List<String> webImports = [
-      "import 'dart:html'",
-      'import "dart:html"',
-      //"import 'package:flutter_web",
-      //'import "package:flutter_web',
-    ];
-
-    if (htmlSrc.isNotEmpty) {
-      outputTabController.selectTab('result');
-    } else if (webImports.any((String import) => dartSrc.contains(import))) {
-      outputTabController.selectTab('result');
-    } else {
-      outputTabController.selectTab('console');
-    }
-  }
-
-  /* Future<GistSummary> _createSummary() {
-    return dartSupportServices.getUnusedMappingId().then((UuidContainer id) {
-      _mappingId = id.uuid;
-      SourcesRequest input = new SourcesRequest()
-        ..sources = {
-          'dart': _context.dartSource,
-          'css': _context.cssSource,
-          'html': _context.htmlSource
-        };
-      return dartServices.summarize(input);
-    }).then((SummaryText summary) {
-      return new GistSummary(summary.text,
-          'Find this at [dartpad.dartlang.org/?source=${_mappingId}](https://dartpad.dartlang.org/?source=${_mappingId}).');
-    }).catchError((e) {
-      _logger.severe(e);
-    });
-  } */
 
   /// Perform static analysis of the source code. Return whether the code
   /// analyzed cleanly (had no errors or warnings).
@@ -752,18 +821,17 @@ class Playground implements GistContainer, GistController {
       bool hasErrors = result.issues.any((issue) => issue.kind == 'error');
       bool hasWarnings = result.issues.any((issue) => issue.kind == 'warning');
 
-      _updateRunButton(hasErrors: hasErrors, hasWarnings: hasWarnings);
+      // TODO: show errors or warnings
 
       return hasErrors == false && hasWarnings == false;
     }).catchError((e) {
       _context.dartDocument.setAnnotations([]);
       busyLight.reset();
-      _updateRunButton();
       _logger.severe(e);
     });
   }
 
-  Future _format() {
+  Future<void> _format() {
     String originalSource = _context.dartSource;
     SourceRequest input = SourceRequest()..source = originalSource;
     formatButton.disabled = true;
@@ -781,9 +849,9 @@ class Playground implements GistContainer, GistController {
 
       if (originalSource != result.newString) {
         editor.document.updateValue(result.newString);
-        DToast.showMessage('Format successful.');
+        _showSnackbar('Format successful.');
       } else {
-        DToast.showMessage('No formatting changes.');
+        _showSnackbar('No formatting changes.');
       }
     }).catchError((e) {
       busyLight.reset();
@@ -795,143 +863,156 @@ class Playground implements GistContainer, GistController {
   void _handleSave() => ga.sendEvent('main', 'save');
 
   void _clearOutput() {
-    _outputpanel.text = '';
+    _rightConsole.clear();
+    _leftConsole.clear();
+    unreadConsoleCounter.clear();
   }
-
-  InputElement get showWebContentInputElement =>
-      querySelector('#show-web-content');
-
-  bool get isWebContentShowing =>
-      querySelector('#htmltab').style.visibility != 'hidden';
-
-  void showWebContent() {
-    showWebContentInputElement.checked = true;
-
-    querySelector('#htmltab').style.visibility = 'initial';
-    querySelector('#csstab').style.visibility = 'initial';
-
-    querySelector('#resulttab').style.visibility = 'initial';
-  }
-
-  void hideWebContent() {
-    showWebContentInputElement.checked = false;
-
-    // Switch to dart tab, and the console output tab.
-    sourceTabController.selectTab('dart');
-    outputTabController.selectTab('console');
-
-    querySelector('#htmltab').style.visibility = 'hidden';
-    querySelector('#csstab').style.visibility = 'hidden';
-
-    querySelector('#resulttab').style.visibility = 'hidden';
-  }
-
-  final _bufferedOutput = <SpanElement>[];
-  final _outputDuration = Duration(milliseconds: 32);
 
   void _showOutput(String message, {bool error = false}) {
-    consoleBusyLight.flash();
+    _leftConsole.showOutput(message, error: error);
+    _rightConsole.showOutput(message, error: error);
 
-    SpanElement span = SpanElement()..text = '$message\n';
-    span.classes.add(error ? 'error-output' : 'normal');
-
-    // Buffer the console output so that heavy writing to stdout does not starve
-    // the DOM thread.
-    _bufferedOutput.add(span);
-
-    if (_bufferedOutput.length == 1) {
-      Timer(_outputDuration, () {
-        _outputpanel.children.addAll(_bufferedOutput);
-        _outputpanel.children.last.scrollIntoView(ScrollAlignment.BOTTOM);
-        _bufferedOutput.clear();
-      });
+    // If there's no tabs visible or the console is not being displayed,
+    // increment the counter
+    if (tabExpandController == null ||
+        tabExpandController?.state != TabState.console) {
+      unreadConsoleCounter.increment();
     }
   }
 
-  void _handleSelectChanged(SelectElement select) {
-    String value = select.value;
+  void _showSnackbar(String message) {
+    var div = querySelector('.mdc-snackbar');
+    var snackbar = MDCSnackbar(div)..labelText = message;
+    snackbar.open();
+  }
 
-    if (isLegalGistId(value)) {
-      router.go('gist', {'gist': value});
+  Layout _detectLayout(Gist gist) {
+    if (gist.hasWebContent()) {
+      return Layout.html;
+    } else if (gist.hasFlutterContent()) {
+      return Layout.flutter;
+    } else {
+      return Layout.dart;
+    }
+  }
+
+  void _changeLayout(Layout layout) {
+    if (_layout == layout) {
+      return;
     }
 
-    select.value = '0';
+    _layout = layout;
+
+    if (layout == Layout.dart) {
+      _frame.hidden = true;
+      editorPanelFooter.setAttr('hidden');
+      _disposeOutputPanelTabs();
+      _rightDocPanel.attributes.remove('hidden');
+      _rightConsoleElement.attributes.remove('hidden');
+      webTabBar.setAttr('hidden');
+      webLayoutTabController.selectTab('dart');
+      _initRightSplitter();
+      editorPanelHeader.setAttr('hidden');
+      webOutputLabel.setAttr('hidden');
+    } else if (layout == Layout.html) {
+      _disposeRightSplitter();
+      _frame.hidden = false;
+      editorPanelFooter.clearAttr('hidden');
+      _initOutputPanelTabs();
+      _rightDocPanel.setAttribute('hidden', '');
+      _rightConsoleElement.setAttribute('hidden', '');
+      webTabBar.toggleAttr('hidden', false);
+      webLayoutTabController.selectTab('dart');
+      editorPanelHeader.clearAttr('hidden');
+      webOutputLabel.setAttr('hidden');
+    } else if (layout == Layout.flutter) {
+      _disposeRightSplitter();
+      _frame.hidden = false;
+      editorPanelFooter.clearAttr('hidden');
+      _initOutputPanelTabs();
+      _rightDocPanel.setAttribute('hidden', '');
+      _rightConsoleElement.setAttribute('hidden', '');
+      webTabBar.setAttr('hidden');
+      webLayoutTabController.selectTab('dart');
+      editorPanelHeader.setAttr('hidden');
+      webOutputLabel.clearAttr('hidden');
+    }
+  }
+
+  // GistContainer interface
+  @override
+  MutableGist get mutableGist => editableGist;
+
+  @override
+  void overrideNextRoute(Gist gist) {
+    _overrideNextRouteGist = gist;
+  }
+
+  Future<void> _showCreateGistDialog() async {
+    var result = await dialog.showOkCancel(
+        'Create New Pad', 'Discard changes to the current pad?');
+    if (result == DialogResult.ok) {
+      var layout = await newPadDialog.show();
+      await createGistForLayout(layout);
+      _changeLayout(layout);
+    }
+  }
+
+  Future<void> _showResetDialog() async {
+    var result = await dialog.showOkCancel(
+        'Reset Pad', 'Discard changes to the current pad?');
+    if (result == DialogResult.ok) {
+      _resetGists();
+    }
+  }
+
+  void _showSharingPage() {
+    window.open('https://github.com/dart-lang/dart-pad/wiki/Sharing-Guide',
+        'DartPad Sharing Guide');
+  }
+
+  void _showGitHubPage() {
+    window.open('https://github.com/dart-lang/dart-pad', 'DartPad on GitHub');
+  }
+
+  @override
+  Future<void> createNewGist() async {
+    _gistStorage.clearStoredGist();
+
+    if (ga != null) ga.sendEvent('main', 'new');
+
+    _showSnackbar('New pad created');
+    await router.go('gist', {'gist': ''}, forceReload: true);
+  }
+
+  Future<void> createGistForLayout(Layout layout) async {
+    _gistStorage.clearStoredGist();
+
+    if (ga != null) ga.sendEvent('main', 'new');
+
+    _showSnackbar('New pad created');
+
+    var layoutStr = _layoutToString(layout);
+
+    await router.go(layoutStr, {}, forceReload: true);
+  }
+
+  void _resetGists() {
+    _gistStorage.clearStoredGist();
+    editableGist.reset();
+    // Delay to give time for the model change event to propagate through
+    // to the editor component (which is where `_performAnalysis()` pulls
+    // the Dart source from).
+    Timer.run(_performAnalysis);
+    _clearOutput();
   }
 
   void _displayIssues(List<AnalysisIssue> issues) {
-    Element issuesElement = querySelector('#issues');
-
-    // Detect when hiding; don't remove the content until hidden.
-    bool isHiding = issuesElement.children.isNotEmpty && issues.isEmpty;
-
-    if (isHiding) {
-      issuesElement.classes.toggle('showing', issues.isNotEmpty);
-
-      StreamSubscription sub;
-      sub = issuesElement.onTransitionEnd.listen((_) {
-        issuesElement.children.clear();
-        sub.cancel();
-      });
-    } else {
-      issuesElement.children.clear();
-
-      issues.sort((a, b) => a.charStart - b.charStart);
-
-      // Create an item for each issue.
-      for (AnalysisIssue issue in issues) {
-        DivElement e = DivElement();
-        e.classes.add('issue');
-        e.attributes['layout'] = '';
-        e.attributes['horizontal'] = '';
-        issuesElement.children.add(e);
-        e.onClick.listen((_) {
-          _jumpTo(issue.line, issue.charStart, issue.charLength, focus: true);
-        });
-
-        SpanElement typeSpan = SpanElement();
-        typeSpan.classes.addAll([issue.kind, 'issuelabel']);
-        typeSpan.text = issue.kind;
-        e.children.add(typeSpan);
-
-        SpanElement messageSpan = SpanElement();
-        messageSpan.classes.add('message');
-        messageSpan.attributes['flex'] = '';
-        messageSpan.text = issue.message;
-        e.children.add(messageSpan);
-        if (issue.hasFixes) {
-          e.classes.add('hasFix');
-          e.onClick.listen((e) {
-            // This is a bit of a hack to make sure quick fixes popup is only
-            // shown if the wrench is clicked, and not if the text or label is
-            // clicked.
-            if ((e.target as Element).className == 'issue hasFix') {
-              // codemiror only shows completions if there is no selected text
-              _jumpTo(issue.line, issue.charStart, 0, focus: true);
-              editor.showCompletions(onlyShowFixes: true);
-            }
-          });
-        }
-      }
-
-      issuesElement.classes.toggle('showing', issues.isNotEmpty);
-    }
-  }
-
-  void _updateRunButton({bool hasErrors = false, bool hasWarnings = false}) {
-    const alertSVGIcon =
-        'M5,3H19A2,2 0 0,1 21,5V19A2,2 0 0,1 19,21H5A2,2 0 0,1 3,19V5A2,2 0 0,'
-        '1 5,3M13,13V7H11V13H13M13,17V15H11V17H13Z';
-
-    var path = runButton.element.querySelector('path');
-    path.attributes['d'] =
-        (hasErrors || hasWarnings) ? alertSVGIcon : 'M8 5v14l11-7z';
-
-    path.parent.classes.toggle('error', hasErrors);
-    path.parent.classes.toggle('warning', hasWarnings && !hasErrors);
+    analysisResultsController.display(issues);
   }
 
   void _jumpTo(int line, int charStart, int charLength, {bool focus = false}) {
-    Document doc = editor.document;
+    final doc = editor.document;
 
     doc.select(
         doc.posFromIndex(charStart), doc.posFromIndex(charStart + charLength));
@@ -940,8 +1021,264 @@ class Playground implements GistContainer, GistController {
   }
 
   void _jumpToLine(int line) {
-    Document doc = editor.document;
+    final doc = editor.document;
     doc.select(Position(line, 0), Position(line, 0));
+
     editor.focus();
   }
+}
+
+enum LoadGistResult {
+  storage,
+  queryParameter,
+  none,
+}
+
+enum Layout {
+  flutter,
+  dart,
+  html,
+}
+
+String _layoutToString(Layout layout) {
+  return layout.toString().split('.').last;
+}
+
+enum TabState {
+  closed,
+  docs,
+  console,
+}
+
+/// Manages the bottom-left panel and tabs
+class TabExpandController {
+  final MDCButton consoleButton;
+  final MDCButton docsButton;
+  final MDCButton closeButton;
+  final DElement console;
+  final DElement docs;
+  final Counter unreadCounter;
+
+  /// The element to give the top half of the split when this panel
+  /// opens
+  final Element topSplit;
+
+  /// The element to give the bottom half of the split
+  final Element bottomSplit;
+
+  final List<StreamSubscription> _subscriptions = [];
+
+  TabState _state;
+  Splitter _splitter;
+  bool _splitterConfigured = false;
+
+  TabState get state => _state;
+
+  TabExpandController({
+    @required this.consoleButton,
+    @required this.docsButton,
+    @required this.closeButton,
+    @required Element consoleElement,
+    @required Element docsElement,
+    @required this.topSplit,
+    @required this.bottomSplit,
+    @required this.unreadCounter,
+  })  : console = DElement(consoleElement),
+        docs = DElement(docsElement) {
+    _state = TabState.closed;
+    console.setAttr('hidden');
+    docs.setAttr('hidden');
+
+    _subscriptions.add(consoleButton.onClick.listen((_) {
+      toggleConsole();
+    }));
+
+    _subscriptions.add(docsButton.onClick.listen((_) {
+      toggleDocs();
+    }));
+
+    _subscriptions.add(closeButton.onClick.listen((_) {
+      _hidePanel();
+    }));
+  }
+
+  void toggleConsole() {
+    if (_state == TabState.closed) {
+      _showConsole();
+    } else if (_state == TabState.docs) {
+      _showConsole();
+      docs.setAttr('hidden');
+      docsButton.toggleClass('active', false);
+    } else if (_state == TabState.console) {
+      _hidePanel();
+    }
+  }
+
+  void toggleDocs() {
+    if (_state == TabState.closed) {
+      _showDocs();
+    } else if (_state == TabState.console) {
+      _showDocs();
+      console.setAttr('hidden');
+      consoleButton.toggleClass('active', false);
+    } else if (_state == TabState.docs) {
+      _hidePanel();
+    }
+  }
+
+  void _showConsole() {
+    unreadCounter.clear();
+    _state = TabState.console;
+    console.clearAttr('hidden');
+    bottomSplit.classes.remove('border-top');
+    consoleButton.toggleClass('active', true);
+    _initSplitter();
+    closeButton.toggleAttr('hidden', false);
+  }
+
+  void _hidePanel() {
+    _destroySplitter();
+    _state = TabState.closed;
+    console.setAttr('hidden');
+    docs.setAttr('hidden');
+    bottomSplit.classes.add('border-top');
+    consoleButton.toggleClass('active', false);
+    docsButton.toggleClass('active', false);
+    closeButton.toggleAttr('hidden', true);
+  }
+
+  void _showDocs() {
+    _state = TabState.docs;
+    docs.clearAttr('hidden');
+    bottomSplit.classes.remove('border-top');
+    docsButton.toggleClass('active', true);
+    _initSplitter();
+    closeButton.toggleAttr('hidden', false);
+  }
+
+  void _initSplitter() {
+    if (_splitterConfigured) {
+      return;
+    }
+
+    _splitter = flexSplit(
+      [topSplit, bottomSplit],
+      horizontal: false,
+      gutterSize: 6,
+      sizes: [70, 30],
+      minSize: [100, 100],
+    );
+    _splitterConfigured = true;
+  }
+
+  void _destroySplitter() {
+    if (!_splitterConfigured) {
+      return;
+    }
+
+    _splitter?.destroy();
+    _splitterConfigured = false;
+  }
+
+  void dispose() {
+    bottomSplit.classes.add('border-top');
+    _destroySplitter();
+
+    // Reset selected tab
+    docsButton.toggleClass('active', false);
+    consoleButton.toggleClass('active', false);
+
+    // Clear listeners
+    _subscriptions.forEach((s) => s.cancel());
+    _subscriptions.clear();
+  }
+}
+
+class NewPadDialog {
+  final MDCDialog _mdcDialog;
+  final MDCRipple _dartButton;
+  final MDCRipple _flutterButton;
+  final MDCButton _createButton;
+  final MDCButton _cancelButton;
+  final MDCSwitch _htmlSwitch;
+  final DElement _htmlSwitchContainer;
+
+  NewPadDialog()
+      : assert(querySelector('#new-pad-dialog') != null),
+        assert(querySelector('#new-pad-select-dart') != null),
+        assert(querySelector('#new-pad-select-flutter') != null),
+        assert(querySelector('#new-pad-cancel-button') != null),
+        assert(querySelector('#new-pad-create-button') != null),
+        assert(querySelector('#new-pad-html-switch-container') != null),
+        assert(querySelector('#new-pad-html-switch-container .mdc-switch') !=
+            null),
+        _mdcDialog = MDCDialog(querySelector('#new-pad-dialog')),
+        _dartButton = MDCRipple(querySelector('#new-pad-select-dart')),
+        _flutterButton = MDCRipple(querySelector('#new-pad-select-flutter')),
+        _cancelButton = MDCButton(querySelector('#new-pad-cancel-button')),
+        _createButton = MDCButton(querySelector('#new-pad-create-button')),
+        _htmlSwitchContainer =
+            DElement(querySelector('#new-pad-html-switch-container')),
+        _htmlSwitch = MDCSwitch(
+            querySelector('#new-pad-html-switch-container .mdc-switch'));
+
+  Layout get selectedLayout {
+    if (_dartButton.root.classes.contains('selected')) {
+      return _htmlSwitch.checked ? Layout.html : Layout.dart;
+    }
+
+    if (_flutterButton.root.classes.contains('selected')) {
+      return Layout.flutter;
+    }
+
+    return null;
+  }
+
+  Future<Layout> show() {
+    _createButton.toggleAttr('disabled', true);
+
+    var completer = Completer<Layout>();
+    var dartSub = _dartButton.root.onClick.listen((_) {
+      _flutterButton.root.classes.remove('selected');
+      _dartButton.root.classes.add('selected');
+      _createButton.toggleAttr('disabled', false);
+      _htmlSwitchContainer.toggleClass('hide', false);
+      _htmlSwitch.disabled = false;
+    });
+
+    var flutterSub = _flutterButton.root.onClick.listen((_) {
+      _dartButton.root.classes.remove('selected');
+      _flutterButton.root.classes.add('selected');
+      _createButton.toggleAttr('disabled', false);
+      _htmlSwitchContainer.toggleClass('hide', true);
+    });
+
+    var cancelSub = _cancelButton.onClick.listen((_) {
+      completer.complete(null);
+    });
+
+    var createSub = _createButton.onClick.listen((_) {
+      completer.complete(selectedLayout);
+    });
+
+    _mdcDialog.open();
+
+    return completer.future.then((v) {
+      _flutterButton.root.classes.remove('selected');
+      _dartButton.root.classes.remove('selected');
+      dartSub.cancel();
+      flutterSub.cancel();
+      cancelSub.cancel();
+      createSub.cancel();
+      _mdcDialog.close();
+      return v;
+    });
+  }
+}
+
+class Sample {
+  final String gistId;
+  final String name;
+  final Layout layout;
+  Sample(this.gistId, this.name, this.layout);
 }
