+++ web/inject_embed.dart
@@ -2,7 +2,7 @@
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
 
-import 'package:dart_pad/experimental/inject_embed.dart' as inject_embed;
+import 'package:dart_pad/inject/inject_embed.dart' as inject_embed;
 
 void main() {
   inject_embed.main();
