+++ web/scripts/main.dart
@@ -2,7 +2,7 @@
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
 
-import 'package:dart_pad/new_playground.dart' as playground;
+import 'package:dart_pad/playground.dart' as playground;
 import 'package:logging/logging.dart';
 
 void main() {
