+++ web/scripts/new_embed_flutter.dart
@@ -2,11 +2,11 @@
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
 
-import 'package:dart_pad/experimental/new_embed.dart';
+import 'package:dart_pad/embed.dart';
 import 'package:logging/logging.dart';
 
 void main() {
-  init(NewEmbedOptions(NewEmbedMode.flutter));
+  init(EmbedOptions(EmbedMode.flutter));
 
   Logger.root.onRecord.listen(print);
 }
